Add SBAS tests
